requirements.txt
flask
flask-cors
flask-sqlalchemy
flasgger
requests
pillow